rag_engine/requirements.txt
requests
pymupdf
yfinance
pyarrow
pinecone>=3.0.0
sentence-transformers>=2.6.0
torch>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.1
langchain-text-splitters>=0.2.0
